# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Examination & OSM Platform",
    "description": "Backend API for AI-assisted Digital Examination & On-Screen Marking Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/exams": {
      "get": {
        "tags": [
          "Exams"
        ],
        "summary": "List Exams Endpoint",
        "description": "E2: List exams. Admins and moderators see all; examiners see exams with assigned answers.",
        "operationId": "list_exams_endpoint_api_v1_exams_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExamResponse"
                  },
                  "type": "array",
                  "title": "Response List Exams Endpoint Api V1 Exams Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Exams"
        ],
        "summary": "Create Exam Endpoint",
        "description": "E1: Create a new exam in draft status.",
        "operationId": "create_exam_endpoint_api_v1_exams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/exams/{id}": {
      "get": {
        "tags": [
          "Exams"
        ],
        "summary": "Get Exam Detail Endpoint",
        "description": "E3: Get exam details with question, sheet, and answer counts.",
        "operationId": "get_exam_detail_endpoint_api_v1_exams__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Exams"
        ],
        "summary": "Update Exam Endpoint",
        "description": "E4: Update exam details or settings.",
        "operationId": "update_exam_endpoint_api_v1_exams__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{id}/transition": {
      "post": {
        "tags": [
          "Exams"
        ],
        "summary": "Transition Exam Endpoint",
        "description": "E5: Transition exam lifecycle state (draft -> evaluation -> moderation -> completed).",
        "operationId": "transition_exam_endpoint_api_v1_exams__id__transition_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{id}/questions": {
      "post": {
        "tags": [
          "Exams"
        ],
        "summary": "Create Question Endpoint",
        "description": "Q1: Add a question to an exam in draft status.",
        "operationId": "create_question_endpoint_api_v1_exams__id__questions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Exams"
        ],
        "summary": "Get Questions Endpoint",
        "description": "Q2: List questions with rubrics for an exam.",
        "operationId": "get_questions_endpoint_api_v1_exams__id__questions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponse"
                  },
                  "title": "Response Get Questions Endpoint Api V1 Exams  Id  Questions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/answers/{id}/ai-evaluation": {
      "post": {
        "tags": [
          "AI Evaluation"
        ],
        "summary": "Trigger Ai Evaluation",
        "description": "V1: Run AI evaluation for one answer (synchronous). Admin or the assigned examiner.\nMode comes from the question unless force_standard is set.",
        "operationId": "trigger_ai_evaluation_api_v1_answers__id__ai_evaluation_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAIEvaluationRequest",
                "default": {
                  "force_standard": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/answers/{id}/ai-evaluations": {
      "get": {
        "tags": [
          "AI Evaluation"
        ],
        "summary": "Get Ai Evaluation History",
        "description": "V3: History of AI suggestions for one answer, newest first.",
        "operationId": "get_ai_evaluation_history_api_v1_answers__id__ai_evaluations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIEvaluationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{id}/ai-evaluation/run": {
      "post": {
        "tags": [
          "AI Evaluation"
        ],
        "summary": "Trigger Batch Ai Evaluation",
        "description": "V2: Batch AI on answers with ocr_status=done and no current suggestion.",
        "operationId": "trigger_batch_ai_evaluation_api_v1_exams__id__ai_evaluation_run_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRunAIEvaluationRequest",
                "default": {
                  "answer_ids": [],
                  "only_pending": true
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchRunAIEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIEvaluationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AIEvaluationResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AIEvaluationListResponse"
      },
      "AIEvaluationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "answer_id": {
            "type": "integer",
            "title": "Answer Id"
          },
          "mode_requested": {
            "type": "string",
            "title": "Mode Requested"
          },
          "mode_used": {
            "type": "string",
            "title": "Mode Used"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "suggested_marks": {
            "type": "number",
            "title": "Suggested Marks"
          },
          "max_marks": {
            "type": "number",
            "title": "Max Marks"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "llm_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Confidence"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionResult"
            },
            "type": "array",
            "title": "Criteria"
          },
          "overall_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Reason"
          },
          "retrieval": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieval"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "answer_id",
          "mode_requested",
          "mode_used",
          "model_name",
          "prompt_version",
          "suggested_marks",
          "max_marks",
          "confidence",
          "criteria"
        ],
        "title": "AIEvaluationResponse"
      },
...
      "BatchRunAIEvaluationRequest": {
        "properties": {
          "answer_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Ids"
          },
          "only_pending": {
            "type": "boolean",
            "title": "Only Pending",
            "default": true
          }
        },
        "type": "object",
        "title": "BatchRunAIEvaluationRequest"
      },
      "BatchRunAIEvaluationResponse": {
        "properties": {
          "queued": {
            "type": "integer",
            "title": "Queued"
          }
        },
        "type": "object",
        "required": [
          "queued"
        ],
        "title": "BatchRunAIEvaluationResponse"
      },
...
      "CriterionResult": {
        "properties": {
          "criterion_id": {
            "type": "string",
            "title": "Criterion Id"
          },
          "criterion": {
            "type": "string",
            "title": "Criterion"
          },
          "max_marks": {
            "type": "number",
            "title": "Max Marks"
          },
          "awarded_marks": {
            "type": "number",
            "title": "Awarded Marks"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "criterion_id",
          "criterion",
          "max_marks",
          "awarded_marks",
          "reason"
        ],
        "title": "CriterionResult"
      },
...
      "ExamCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "course_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Course Code"
          },
          "exam_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Date"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings"
          }
        },
        "type": "object",
        "required": [
          "title",
          "course_code"
        ],
        "title": "ExamCreate"
      },
      "ExamDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "exam_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "evaluation",
              "moderation",
              "completed"
            ],
            "title": "Status"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "default": 0
          },
          "sheet_count": {
            "type": "integer",
            "title": "Sheet Count",
            "default": 0
          },
          "answer_count": {
            "type": "integer",
            "title": "Answer Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "course_code",
          "status",
          "settings",
          "created_at"
        ],
        "title": "ExamDetailResponse"
      },
      "ExamResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "course_code": {
            "type": "string",
            "title": "Course Code"
          },
          "exam_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Date"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "evaluation",
              "moderation",
              "completed"
            ],
            "title": "Status"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "course_code",
          "status",
          "settings",
          "created_at"
        ],
        "title": "ExamResponse"
      },
      "ExamTransitionRequest": {
        "properties": {
          "to": {
            "type": "string",
            "enum": [
              "evaluation",
              "moderation",
              "completed"
            ],
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "to"
        ],
        "title": "ExamTransitionRequest"
      },
      "ExamUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "course_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "exam_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exam Date"
          },
          "settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings"
          }
        },
        "type": "object",
        "title": "ExamUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionCreate": {
        "properties": {
          "question_number": {
            "type": "string",
            "maxLength": 10,
            "minLength": 1,
            "title": "Question Number"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "max_marks": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Max Marks"
          },
          "evaluation_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "standard",
                  "reference_grounded"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Mode",
            "default": "standard"
          },
          "display_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "question_number",
          "text",
          "max_marks"
        ],
        "title": "QuestionCreate"
      },
      "QuestionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exam_id": {
            "type": "integer",
            "title": "Exam Id"
          },
          "question_number": {
            "type": "string",
            "title": "Question Number"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "max_marks": {
            "type": "number",
            "title": "Max Marks"
          },
          "evaluation_mode": {
            "type": "string",
            "enum": [
              "standard",
              "reference_grounded"
            ],
            "title": "Evaluation Mode"
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order"
          },
          "rubric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RubricResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "exam_id",
          "question_number",
          "text",
          "max_marks",
          "evaluation_mode",
          "display_order"
        ],
        "title": "QuestionResponse"
      },
...
      "RubricResponse": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/RubricCriterion"
            },
            "type": "array",
            "title": "Criteria"
          },
          "guidance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guidance"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "criteria"
        ],
        "title": "RubricResponse"
      },
      "RunAIEvaluationRequest": {
        "properties": {
          "force_standard": {
            "type": "boolean",
            "title": "Force Standard",
            "description": "Skip reference retrieval for a reference_grounded question. Logged as mode_requested='standard'.",
            "default": false
          }
        },
        "type": "object",
        "title": "RunAIEvaluationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}